WTA #71: Moved common dto to single file.
