+++ core/common/src/main/java/com/jacob/wakatimeapp/core/common/data/dtos/CoreDTOs.kt
@@ -1,0 +1,60 @@
+package com.jacob.wakatimeapp.core.common.data.dtos
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+class DependencyDTO
+
+@Serializable
+data class EditorDTO(
+    val decimal: String,
+    val digital: String,
+    val hours: Int,
+    val minutes: Int,
+    val name: String,
+    val percent: Double,
+    val seconds: Int,
+    val text: String,
+    @SerialName("total_seconds") val totalSeconds: Double,
+)
+
+@Serializable
+data class LanguageDTO(
+    val decimal: String,
+    val digital: String,
+    val hours: Int,
+    val minutes: Int,
+    val name: String,
+    val percent: Double,
+    val seconds: Int,
+    val text: String,
+    @SerialName("total_seconds") val totalSeconds: Double,
+)
+
+@Serializable
+data class MachineDTO(
+    val decimal: String,
+    val digital: String,
+    val hours: Int,
+    val minutes: Int,
+    val name: String,
+    val percent: Double,
+    val seconds: Int,
+    val text: String,
+    @SerialName("machine_name_id") val machineNameId: String,
+    @SerialName("total_seconds") val totalSeconds: Double,
+)
+
+@Serializable
+data class OperatingSystemDTO(
+    val decimal: String,
+    val digital: String,
+    val hours: Int,
+    val minutes: Int,
+    val name: String,
+    val percent: Double,
+    val seconds: Int,
+    val text: String,
+    @SerialName("total_seconds") val totalSeconds: Double,
+)
