--- core/common/src/main/java/com/jacob/wakatimeapp/core/common/data/dtos/DependencyDTO.kt
@@ -1,6 +1,0 @@
-package com.jacob.wakatimeapp.core.common.data.dtos
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-class DependencyDTO
