--- core/common/src/main/java/com/jacob/wakatimeapp/core/common/data/dtos/LanguageDTO.kt
@@ -1,17 +1,0 @@
-package com.jacob.wakatimeapp.core.common.data.dtos
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class LanguageDTO(
-    val decimal: String,
-    val digital: String,
-    val hours: Int,
-    val minutes: Int,
-    val name: String,
-    val percent: Double,
-    val seconds: Int,
-    val text: String,
-    @SerialName("total_seconds") val totalSeconds: Double,
-)
